Fixed the conflict issue with Employee Appraisal

+++ src/main/java/com/asg/payroll/salarydetails/dto/SalaryDetailRequest.java
@@ -18,8 +18,8 @@ public class SalaryDetailRequest {
     private LocalDate nextIncrementDate;
     private LocalDate indemnityDueFrom;
     
-    private Long basicSalary;
-    private Long registeredSalary; // Gosi Salary
+    private BigDecimal basicSalary;
+    private BigDecimal registeredSalary; // Gosi Salary
     private String gosiType;
     
     private String paymentMethod;
@@ -33,10 +33,10 @@ public class SalaryDetailRequest {
     private LocalDate bankGuaranteeDueDate;
 
     private String contractPrintType;
-    private Long indemnityPaidAmt;
+    private BigDecimal indemnityPaidAmt;
     private String indemnityPaidDet;
-    private Long loanDeductionAmt;
-    private Long accommodationCost;
+    private BigDecimal loanDeductionAmt;
+    private BigDecimal accommodationCost;
     private String remarks;
 
     private List<SalaryAllowanceDto> allowances;
